rewrite AutoBroadcast and update CaiBot
