+++ src/AutoBroadcast/AutoBroadcast.cs
@@ -8,13 +8,14 @@ namespace AutoBroadcast;
 [ApiVersion(2, 1)]
 public class AutoBroadcast : LazyPlugin
 {
-    public override string Name => System.Reflection.Assembly.GetExecutingAssembly().GetName().Name!; public override string Author => "Scavenger";
+    public override string Name => System.Reflection.Assembly.GetExecutingAssembly().GetName().Name!;
+    public override string Author => "Scavenger,Cai";
     public override string Description => GetString("自动广播插件");
-    public override Version Version => new Version(1, 0, 8);
+    public override Version Version => new (1, 1, 1);
 
-    public DateTime LastCheck = DateTime.UtcNow;
+    private DateTime _lastCheck = DateTime.UtcNow;
 
-    public AutoBroadcast(Main Game) : base(Game) { }
+    public AutoBroadcast(Main game) : base(game) { }
 
     public override void Initialize()
     {
@@ -22,167 +23,64 @@ public class AutoBroadcast : LazyPlugin
         ServerApi.Hooks.ServerChat.Register(this, this.OnChat);
     }
 
-    protected override void Dispose(bool Disposing)
+    protected override void Dispose(bool disposing)
     {
-        if (Disposing)
+        if (disposing)
         {
             ServerApi.Hooks.GameUpdate.Deregister(this, this.OnUpdate);
             ServerApi.Hooks.ServerChat.Deregister(this, this.OnChat);
         }
-        base.Dispose(Disposing);
+        base.Dispose(disposing);
     }
+    
+    public void OnUpdate(EventArgs args)
+    {
+        if (!((DateTime.UtcNow - this._lastCheck).TotalSeconds >= 1))
+        {
+            return;
+        }
+        this._lastCheck = DateTime.UtcNow;
+        
+        foreach (var broadcast in AutoBroadcastConfig.Instance.Broadcasts)
+        {
+            if (!broadcast.Enabled)
+            {
+                continue;
+            }
 
-
-
-    #region Chat
+            broadcast.SecondUpdate();
+        }
+    }
     public void OnChat(ServerChatEventArgs args)
     {
         if (TShock.Players[args.Who] == null)
         {
             return;
         }
-        var Groups = Array.Empty<string>();
-        var Messages = Array.Empty<string>();
-        var Colour = Array.Empty<float>();
-        var PlayerGroup = TShock.Players[args.Who].Group.Name;
 
-        lock (ABConfig.Instance.Broadcasts)
+        var plr = TShock.Players[args.Who];
+        
+        foreach (var broadcast in AutoBroadcastConfig.Instance.Broadcasts)
         {
-            foreach (var broadcast in ABConfig.Instance.Broadcasts)
+            if (!broadcast.Enabled)
             {
-                if (broadcast == null || !broadcast.Enabled ||
-                   (broadcast.TriggerToWholeGroup && !broadcast.Groups.Contains(PlayerGroup)))
-                {
-                    continue;
-                }
-
-                foreach (var Word in broadcast.TriggerWords)
-                {
-                    if (args.Text.Contains(Word))
-                    {
-                        if (broadcast.TriggerToWholeGroup && broadcast.Groups.Length > 0)
-                        {
-                            Groups = broadcast.Groups;
-                        }
-                        Messages = broadcast.Messages;
-                        Colour = broadcast.ColorRGB;
-                        break;
-                    }
-                }
-            }
-        }
-
-        if (Groups.Length > 0)
-        {
-            BroadcastToGroups(Groups, Messages, Colour);
-        }
-        else
-        {
-            BroadcastToPlayer(args.Who, Messages, Colour);
-        }
-    }
-    #endregion
-
-    #region Update
-    public void OnUpdate(EventArgs args)
-    {
-        if ((DateTime.UtcNow - this.LastCheck).TotalSeconds >= 1)
-        {
-            this.LastCheck = DateTime.UtcNow;
-            var NumBroadcasts = 0;
-            lock (ABConfig.Instance.Broadcasts)
-            {
-                NumBroadcasts = ABConfig.Instance.Broadcasts.Length;
+                continue;
             }
 
-            for (var i = 0; i < NumBroadcasts; i++)
+            if (!broadcast.Groups.Contains(plr.Group.Name))
             {
-                var Groups = Array.Empty<string>();
-                var Messages = Array.Empty<string>();
-                var Colour = Array.Empty<float>();
-
-                lock (ABConfig.Instance.Broadcasts)
+                continue;
+            }
+            
+            foreach (var word in broadcast.TriggerWords)
+            {
+                if (args.Text.Contains(word))
                 {
-                    if (ABConfig.Instance.Broadcasts[i] == null || ABConfig.Instance.Broadcasts[i].Enabled || ABConfig.Instance.Broadcasts[i].Interval < 1)
-                    {
-                        continue;
-                    }
-                    if (ABConfig.Instance.Broadcasts[i].StartDelay > 0)
-                    {
-                        ABConfig.Instance.Broadcasts[i].StartDelay--;
-                        continue;
-                    }
-                    ABConfig.Instance.Broadcasts[i].StartDelay = ABConfig.Instance.Broadcasts[i].Interval;// Start Delay used as Interval Countdown
-                    Groups = ABConfig.Instance.Broadcasts[i].Groups;
-                    Messages = ABConfig.Instance.Broadcasts[i].Messages;
-                    Colour = ABConfig.Instance.Broadcasts[i].ColorRGB;
-                }
-
-                if (Groups.Length > 0)
-                {
-                    BroadcastToGroups(Groups, Messages, Colour);
-                }
-                else
-                {
-                    BroadcastToAll(Messages, Colour);
+                    broadcast.RunTriggerWords(plr);
+                    break;
                 }
             }
         }
     }
-    #endregion
-
-    public static void BroadcastToGroups(string[] Groups, string[] Messages, float[] Colour)
-    {
-        foreach (var Line in Messages)
-        {
-            if (Line.StartsWith("/"))
-            {
-                Commands.HandleCommand(TSPlayer.Server, Line);
-            }
-            else
-            {
-                lock (TShock.Players)
-                {
-                    foreach (var player in TShock.Players)
-                    {
-                        if (player != null && Groups.Contains(player.Group.Name))
-                        {
-                            player.SendMessage(Line, (byte) Colour[0], (byte) Colour[1], (byte) Colour[2]);
-                        }
-                    }
-                }
-            }
-        }
-    }
-    public static void BroadcastToAll(string[] Messages, float[] Colour)
-    {
-        foreach (var Line in Messages)
-        {
-            if (Line.StartsWith("/"))
-            {
-                Commands.HandleCommand(TSPlayer.Server, Line);
-            }
-            else
-            {
-                TSPlayer.All.SendMessage(Line, (byte) Colour[0], (byte) Colour[1], (byte) Colour[2]);
-            }
-        }
-    }
-    public static void BroadcastToPlayer(int plr, string[] Messages, float[] Colour)
-    {
-        foreach (var Line in Messages)
-        {
-            if (Line.StartsWith("/"))
-            {
-                Commands.HandleCommand(TSPlayer.Server, Line);
-            }
-            else
-            {
-                lock (TShock.Players)
-                {
-                    TShock.Players[plr].SendMessage(Line, (byte) Colour[0], (byte) Colour[1], (byte) Colour[2]);
-                }
-            }
-        }
-    }
+    
 }
