+++ src/AutoBroadcast/README.md
@@ -22,9 +22,9 @@
         "自动广播执行了服务器指令/say Ciallo～(∠・ω< )⌒★"
       ],
       "RGB颜色": [
-        255.0,
-        234.0,
-        115.0
+        255,
+        234,
+        115
       ],
       "时间间隔": 600,
       "延迟执行": 0,
@@ -39,6 +39,8 @@
 ## 更新日志
 
 ```
+v1.1.1
+几乎完全重写逻辑
 v1.0.8
 准备更新TS 5.2.1,修正文档，初始配置内容更改
 v1.0.6
